Moved ai into its own folder
